requirements.txt
accelerate==0.25.0
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
av==12.3.0
babel==2.16.0
boto3==1.35.68
botocore==1.35.68
cached_path==1.6.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.1.8
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
einops==0.7.0
einops-exts==0.0.4
executing==2.1.0
filelock==3.12.4
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.73.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.0
grpcio-status==1.68.0
gruut==2.4.0
gruut-ipa==0.13.0
gruut-lang-en==2.0.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.19.4
humanfriendly==10.0
icecream==2.1.3
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.20
langchain-google-genai==2.0.5
langchain-google-vertexai==2.0.7
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.144
lazy_loader==0.4
librosa==0.10.2.post1
livekit==0.18.0
livekit-agents==0.11.1
livekit-api==0.7.1
livekit-plugins-openai==0.10.5
livekit-plugins-silero==0.7.3
livekit-protocol==0.6.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munch==4.0.0
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.20.0
openai==1.54.4
opencv-python==4.10.0.84
orjson==3.10.11
outcome==1.3.0.post0
packaging==23.2
pillow==10.3.0
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.9.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.9.0
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.26.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.11.0
SQLAlchemy==2.0.35
sseclient-py==1.8.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.0
transformers==4.40.2
trio==0.27.0
trio-websocket==0.11.1
triton==3.1.0
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.3
watchfiles==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
yarl==1.17.1
